# [PyTorch] Add int version of vectorized PrefixSum to Benchmark